handle big uploaded files
TemporaryUploadedFiles get created when a big file (bigger than 2.5MB by
default) is uploaded.  When the file gets saved to the filesystem, Django
optimizes this by moving the temporary file.  However, when PIL tries to open
it again, the file isn't there and we get an IOError.  To work around this, we
reopen the file if we had a temporary file first.