Bug #12492: Submit thanks page links to wrong video
Previously, the thanks page would just link to the video with the highest id
submitted by the user.  However, this broke down for cases where we were
approving an older video.  Now, the thanks page optionally takes the ID of a
video ID, and uses that to figure out which video to link to.