Adjust tests to hard-code "testserver" everywhere :|
This makes a little bit of sense -- the site.domain is what we prefer to use,
but redirects always use whatever the browser in question sent us. The thing
is that the test client always uses "testserver" as its HTTP_HOST value,
so redirects we generate will also use that.