Mail that we used to send in the SiteLocation.pre_save() hook: Now we send in post_save
This change decreases the amount of time the database object save() takes. It is
important to keep that time low so that two different simultaneous processes do not
try to read/write the SiteLocation and be confused as to the state.

In theory, we should fix this with more locking, honestly, or stronger use of
database transactions. It is wise to do the locking change as well -- even if we
do that, this change decreases the amount of time the process has to hold the lock.