Create CHANGE_STAMPS functionality
Many sites have no writes for a while. To accelerate running management tasks
we can save a note on disk with the timestamp of the last time that e.g.
update_index was run. If update_index was run more recently than the video
updated stamp time, we know it is safe to not run update_index.