trust-but-verify the video counts
Despite using `.update()`, it's very possible to have the skip or import counts
be off-by-one, or more.  Before, this would keep the import from finishing
since mark_import_complete would never think the import was done.  Now, we
calculate the real counts from the ImportIndexes in the database, which are not
prone to race conditions.