don't raise exceptions on 404ed thumbnails
Previously, we were raising CannotOpenImageUrl for this case.  However, given
the large number of 404ed thumbnails I see in the log, it's hardly a fatal
issue.  We've already got an command to go through and try to re-fetch
thumbnails, so we piggyback on that to periodically try to get them.  In the
meantime, we set the `has_thumbnail` flag to False so that the thumbnailing
code will fall back to the feed thumbnail if it's available.