I give up; going with the easy fix
Something, somewhere, isn't quite resetting the Tier state.  What happens is
that when we get to this TestCase, we're enforcing the tier, which we shouldn't
be doing. Since the default tier is 'basic', this caused us to remove the admin
account when we saved the settings.  Setting the tier to 'max' should resolve
the symptom of the failing test case.  Hopefully, when we factor out the Tier
stuff into a separate app, this problem will go away.