implement `watch_count` as an Aggregate
MySQL doesn't allow statements in the WHERE section to refer to computed values
like `watch_count`.  So, we need to put it in a HAVING section, and
`QuerySet.extra()` doesn't allow that.  Implementing `WatchCount()` as an
Aggregate tells Django that it can refer to it, but it needs to be in HAVING.