refactor PlaylistView tests into units
I didn't touch the others; it's not really simpler to make requests manually
vs. using the test client.  The test client also allows us to make assertions
about the templates, which direct requests do not.

This also reverts the permanent redirect from incorrect URLs in playlists; it's
back to being a 302 along with a note that it's explicitly a 302.